fix: splitting the utils for avoiding calling resource packs staffs too early
fix #2
