+++ build.gradle
@@ -1,6 +1,6 @@
 plugins {
     id "architectury-plugin" version "3.4-SNAPSHOT"
-    id "dev.architectury.loom" version "1.2-SNAPSHOT" apply false
+    id "dev.architectury.loom" version "1.3.+" apply false
     id 'io.github.juuxel.loom-vineflower' version '1.11.0' apply false
 }
