+++ common/build.gradle
@@ -10,7 +10,7 @@ dependencies {
     // We depend on fabric loader here to use the fabric @Environment annotations and get the mixin dependencies
     // Do NOT use other classes from fabric loader
     modImplementation "net.fabricmc:fabric-loader:${rootProject.fabric_loader_version}"
-    include(implementation("blue.endless:jankson:${rootProject.jankson_version}"))
+    implementation("blue.endless:jankson:${rootProject.jankson_version}")
 
     modImplementation "com.github.glitchfiend:TerraBlender-fabric:${rootProject.minecraft_version}-${rootProject.terrablender_version}"
 }
