+++ common/src/main/java/net/cristellib/mixin/MixinTerrablenderConfig.java
@@ -7,7 +7,7 @@ import org.spongepowered.asm.mixin.injection.Redirect;
 import terrablender.config.Config;
 import terrablender.config.TerraBlenderConfig;
 
-@Mixin(TerraBlenderConfig.class)
+@Mixin(value = TerraBlenderConfig.class, remap = false)
 public class MixinTerrablenderConfig {
     @Redirect(method = "<init>", at = @At(value = "INVOKE", target = "Lterrablender/config/Config;addNumber(Ljava/lang/String;Ljava/lang/String;Ljava/lang/Number;Ljava/lang/Number;Ljava/lang/Number;)Ljava/lang/Number;"))
     private <T extends Number & Comparable<T>> Number create(Config instance, String comment, String key, T defaultValue, T min, T max) {
