+++ gradle.properties
@@ -12,8 +12,8 @@ fabric_api_version=0.89.2+1.20.2
 
 forge_version=1.20.2-48.0.13
 
-quilt_loader_version=0.20.2
-quilt_fabric_api_version=7.3.0+0.89.0-1.20.1
+quilt_loader_version=0.21.0
+quilt_fabric_api_version=7.3.1+0.89.3-1.20.1
 
-jankson_version = 1.2.1
+jankson_version=1.2.1
 terrablender_version=3.0.0.170
